move test to files
